changed trade_id column to bigint to handle large ints- fixed bug when inserting thousands of rows a time by adding a batch insert db api- improved semantic error logging

+++ server/data/migrations/20190714220125_bootstrap.js
@@ -3,7 +3,7 @@ exports.up = function (knex) {
         table.primary(['exchange', 'trading_pair', 'time', 'price', 'amount'])
             .unique(['exchange', 'trading_pair', 'time', 'price', 'amount'])
         table.datetime('time')
-        table.integer('trade_id')
+        table.bigInteger('trade_id')
         table.float('price')
         table.float('amount')
         table.string('exchange')
